feat: Comment 수정 및 삭제시 작성자와 요청유저 일치여부 검증하여 작동하도록 기능 추가

+++ core/src/main/java/io/devarium/core/domain/comment/service/CommentServiceImpl.java
@@ -41,12 +41,15 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public Comment updateComment(Long commentId, UpsertComment request, User user) {
         Comment comment = getComment(commentId);
+        comment.validateAuthor(user.getId());
         comment.updateContent(request.content());
         return commentRepository.save(comment);
     }
 
     @Override
     public void deleteComment(Long commentId, User user) {
+        Comment comment = getComment(commentId);
+        comment.validateAuthor(user.getId());
         replyService.deleteRepliesByCommentId(commentId);
         commentRepository.deleteById(commentId);
     }
